pyproject.toml
[tool.poetry]
name = "demo-pyton"
version = "0.1.0"
description = ""
authors = ["Shapran Taras <shaprantaras2000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
mysqlclient = "^2.2.0"
djangorestframework-simplejwt = "^5.3.0"
pillow = "^10.1.0"
django-cors-headers = "^4.3.0"
drf-yasg = "^1.21.7"
celery = "^5.3.6"
redis = "^5.0.1"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.5.0"
djangochannelsrestframework = "^1.2.0"
channels-redis = "^4.1.0"
daphne = "^4.0.0"
django-storages = {extras = ["s3"], version = "^1.14.2"}
boto3 = "^1.34.27"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
django-dotenv = "^1.4.2"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
